add casts to convert integer to double for tab size

+++ src/EditorWidget.cpp
@@ -146,7 +146,7 @@ int EditorWidget::tabWidth() const {
 
 void EditorWidget::setTabWidth(int width) {
   tabWidth_ = width;
-  setTabStopDistance(static_cast<double>(fontMetrics().horizontalAdvance(' ') * width));
+  setTabStopDistance(static_cast<double>(fontMetrics().horizontalAdvance(' ')) * static_cast<double>(width));
 }
 
 int EditorWidget::indentWidth() const {
